hack to switch function for parallel computation based on OS

parallel::mclapply doesn't work on Windows, because forking is not supported. This function defines a socket version of mclapply for windows computer An implementation that switch automatically the parallel process when detecting the os. The code below was inspired from parallel mclapply, Nathan VanHoudnos, Kevin Kuang, Peter Solymos and Eric Archer.

mclapply_win(
  X,
  FUN,
  ...,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = NULL,
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE
)

.stackr_parallel_mc(
  X,
  FUN,
  ...,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = getOption("mc.cores", 2L),
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE,
  affinity.list = NULL
)

.stackr_parallel(
  X,
  FUN,
  ...,
  mc.style = "ETA",
  mc.substyle = NA,
  mc.cores = getOption("mc.cores", 2L),
  ignore.interactive = getOption("ignore.interactive", F),
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE
)

Arguments

X

see parallel mclapply

FUN

see parallel mclapply

...

see parallel mclapply

mc.preschedule

see parallel mclapply

mc.set.seed

see parallel mclapply

mc.silent

see parallel mclapply

mc.cores

see parallel mclapply

mc.cleanup

see parallel mclapply

mc.allow.recursive

see parallel mclapply

Value

For mclapply, a list of the same length as X and named by X.

For mclapply, a list of the same length as X and named by X.