Summary of batch_x.genotypes.txt and batch_x.markers.tsv files.

Useful summary of batch_x.genotypes.txt and batch_x.markers.tsv files produced by STACKS genotypes module use for linkage mapping. Filter segregation distortion and output JoinMap and or OneMap file.

summary_genotypes(
  genotypes,
  markers,
  filter.monomorphic = TRUE,
  filter.missing.band = TRUE,
  filter.mean.log.likelihood = NULL,
  B = NULL,
  filter.GOF = NULL,
  filter.GOF.p.value = NULL,
  ind.genotyped = 1,
  joinmap = NULL,
  onemap = NULL,
  filename = NULL
)

Arguments

genotypes

The genotypes = batch_x.genotypes.txt created by STACKS genotypes module.

markers

The markers = batch_x.markers.tsv created by STACKS genotypes module.

filter.monomorphic

(optional) Should monomorphic loci be filtered out. Default: filter.monomorphic = TRUE.

filter.missing.band

(optional) Should loci with missing band be filtered out. Default: filter.missing.band = TRUE.

filter.mean.log.likelihood

(optional, integer) Apply a mean log likelihood filter to the loci. e.g. filter.mean.log.likelihood = -10. Default: filter.mean.log.likelihood = NULL (no filter)

B

(optional, integer) The segregation distortion p-value will be computed with a Monte Carlo test with B replicates, if B is supplied. For more details, see chisq.test. Default: B = NULL.

filter.GOF

(optional, integer) Filer value of the goodness-of-fit for segregation distortion. Default: filter.GOF = NULL.

filter.GOF.p.value

(optional, double) Filer the goodness-of-fit p-value of GOF segregation distortion. Default: filter.GOF.p.value = NULL.

ind.genotyped

(optional, integer) Filter the number of individual progenies required to keep the marker. Default: ind.genotyped = 1.

joinmap

(optional) Name of the JoinMap file to write in the working directory. e.g. "joinmap.turtle.family3.loc". Default: joinmap = NULL.

onemap

(optional) Name of the OneMap file to write in the working directory. e.g. "onemap.turtle.family3.txt". Default: onemap = NULL.

filename

(optional) The name of the summary file written in the directory. No default. Default: filename = NULL.

Value

The function returns a list with the genotypes summary $geno.sum, joinmap markers $joinmap.sum and onemap markers $onemap.sum summary (use $ to access each components). A JoinMap and/or OneMap file can also be exported to the working direcvtory.

Details

SIGNIFICANCE results pvalue (goodness-of-fit pvalue): <= 0.0001 = ****, <= 0.001 & > 0.0001 = ***, <= 0.01 & > 0.001 = **, <= 0.05 & > 0.01 = *.

work in progress...

References

Catchen JM, Amores A, Hohenlohe PA et al. (2011) Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences. G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013) Stacks: an analysis tool set for population genomics. Molecular Ecology, 22, 3124-3140.

Author

Thierry Gosselin thierrygosselin@icloud.com

Examples

if (FALSE) { # \dontrun{
linkage.map.crab <- summary_genotypes(
genotypes = "batch_10.markers.tsv",
markers = "batch_10.genotypes_300.txt",
filter.monomorphic = TRUE,
filter.missing.band = TRUE,
filter.mean.log.likelihood = -10,
B = NULL,
ind.genotyped = 300,
joinmap = "test.loc",
onemap = "test.onemap.txt",
filename = "genotypes.summary.tsv"
)
} # }