Summarise the reads for indel and GC content and produce the read depth plot

Still in dev, but work nicely. Summarise the reads for indel and GC content and produce the read depth plot

summary_reads(
  fq.files,
  paired.end = FALSE,
  output = "08_stacks_results/02_summary_reads",
  read.depth.plot = TRUE,
  min.coverage.fig = 7L,
  parallel.core = parallel::detectCores() - 1
)

Arguments

fq.files

(character, path) The path to the individual fastq file to check, or the entire fq folder.

paired.end

(logical) Are the files paired-end. Default: paired.end = FALSE.

output

(path) Write the output in a specific directory. Default: output = "08_stacks_results/02_summary_reads".

read.depth.plot

(logical) To get the interesting summaries on the read content, the function is very close to similar to read_depth_plot, so you can produce it for each sample with minimal cost on time. Default: read.depth.plot = TRUE.

min.coverage.fig

(integer). Minimum coverage used to draw the color on the figure. Default: min.coverage.fig = 7L.

parallel.core

(integer) Enable parallel execution with the number of threads. Default: parallel.core = parallel::detectCores() - 1.

Value

The function returns the read depth groups plot and the read stats overall and by read depth groups.

Examples

if (FALSE) { # \dontrun{
require(vroom)
sum <- summary_reads(fq.files = "my_fq_folder")
} # }