Removes the noise of an individual fastq file

This function reads the fastq file of an individual and clean it by removing:

unique reads with high coverage (likely paralogs or TE)

distinct reads with low coverage

clean_fq(
  fq.file,
  min.coverage.threshold = 2L,
  remove.unique.reads = TRUE,
  write.blacklist = FALSE,
  write.blacklist.fasta = FALSE,
  compress = FALSE,
  parallel.core = parallel::detectCores() - 1
)

Arguments

fq.file
(character, path). The path to the individual fastq file to check. Default: fq.file = "my-sample.fq.gz".

min.coverage.threshold
(character, path). Minimum coverage threshold. The function will remove distinct reads with coverage <= to the threshold. Default: min.coverage.threshold = 2L.

remove.unique.reads
(logical). Remove distinct unique reads with high coverage. Likely paralogs or Transposable elements. Default: remove.unique.reads = TRUE.

write.blacklist
(logical). Write the blacklisted reads to a file. Default: write.blacklist = FALSE.

write.blacklist.fasta
(logical). Write the blacklisted reads to a fasta file. Default: write.blacklist.fasta = FALSE.

compress
(logical) To compress the output files. If you have the disk space, don't compress, it's way faster this way to write. Default: compress = FALSE.

parallel.core
(integer) Enable parallel execution with the number of threads. Default: parallel.core = parallel::detectCores() - 1.

Value

The function returns a cleaned fq file with the name of the sample and _cleaned appended to the filename.

Details

coming soon, just try it in the meantime...

Examples

if (FALSE) {
require(vroom)
clean.id <- stackr::clean_fq(
  fq.file = "my-sample.fq.gz",
  min.coverage.threshold = 7,
  remove.unique.reads = TRUE,
  write.blacklist.fasta = TRUE
  )
}